Support for making a CMake build system from evoke. Fixes #6.
* Support for making a CMake build system from evoke

The target of this is not to use cmake as a build system per say, but
to make a cmake structure that IDEs and tooling that only understand
cmake can consume.

It properly creates dependencies using the discovery logic in evoke.
However for system install components it does not do a proper cmake
find_package/....  Instead it creates interface libraries to define
relationships.

A new command line options is introduced '-cm' which when given creates
a set of CMakeList.txt files.

To facilitate the use of global options from the evoke.conf file the
toolsets are extended, providing an api which can take an command line
option string and break it out into a set of compile, include, and link
directives which should be used.  This is stored in a structure which
the CMakeLists building code uses to set global options (per target).

Updated the gitignore list to ignore idea/clion/cmake files.

Added a note to the command line documentation about cmake builds.