Use only hierarchical name of component to search for it
component.name and component.hierarchical_name coincide
in the case of depth 1 components. In the case of
subcomponents, component.name = "comp_subcomp" and
component.hierarchical_name = "comp/subcomp". Only
the later was used to add components to a project,
so there is no need to look for component.name to know
if a component should be tagged as system component.